pyproject.toml
[project]
name = "video-streaming-app"
version = "0.1.0"
description = ""
authors = [
    {name = "SubidShrestha",email = "subideminem@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
